Create common legend test & swap div for <>
