+++ src/components/OSCALCatalogGroup.test.js
@@ -1,6 +1,7 @@
 import React from "react";
 import { render, screen, fireEvent } from "@testing-library/react";
 import OSCALCatalogGroup from "./OSCALCatalogGroup";
+import testOSCALControlParamLegend from "../common-tests/ControlParamLegend.test";
 
 const testGroup = {
   id: "parent-group",
@@ -34,16 +35,7 @@ test("OSCALCatalog displays nested control groups", () => {
   expect(control).toBeVisible();
 });
 
-test("OSCALCatalogGroup displays legend", async () => {
-  render(<OSCALCatalogGroup group={testGroup} />);
-  const parentGroup = screen.getByText("Parent Group");
-  fireEvent.click(parentGroup);
-  const placeholderBox = screen.getByLabelText("legend-placeholder-label");
-  expect(placeholderBox).toBeVisible();
-  const placeholderBoxLabel = screen.getByText("Placeholder");
-  expect(placeholderBoxLabel).toBeVisible();
-  const valueBox = screen.getByLabelText("legend-value-label");
-  expect(valueBox).toBeVisible();
-  const valueBoxLabel = screen.getByText("Value");
-  expect(valueBoxLabel).toBeVisible();
-});
+testOSCALControlParamLegend(
+  "OSCALCatalogGroup",
+  <OSCALCatalogGroup group={testGroup} />
+);
